+++ internal/provider/resource_fmc_ftd_platform_settings_dns_test.go
@@ -31,7 +31,7 @@ import (
 func TestAccFmcFTDPlatformSettingsDNS(t *testing.T) {
 	var checks []resource.TestCheckFunc
 	checks = append(checks, resource.TestCheckResourceAttrSet("fmc_ftd_platform_settings_dns.test", "type"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_ftd_platform_settings_dns.test", "server_groups.0.is_default", "true"))
+	checks = append(checks, resource.TestCheckResourceAttr("fmc_ftd_platform_settings_dns.test", "dns_server_groups.0.is_default", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_ftd_platform_settings_dns.test", "expire_entry_timer", "1"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_ftd_platform_settings_dns.test", "poll_timer", "240"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_ftd_platform_settings_dns.test", "use_management_interface", "true"))
@@ -83,7 +83,7 @@ resource "fmc_dns_server_group" "test" {
 func testAccFmcFTDPlatformSettingsDNSConfig_all() string {
 	config := `resource "fmc_ftd_platform_settings_dns" "test" {` + "\n"
 	config += `	ftd_platform_settings_id = fmc_ftd_platform_settings.test.id` + "\n"
-	config += `	server_groups = [{` + "\n"
+	config += `	dns_server_groups = [{` + "\n"
 	config += `		id = fmc_dns_server_group.test.id` + "\n"
 	config += `		is_default = true` + "\n"
 	config += `	}]` + "\n"